Dialog: Move dialog into widgets folder
